+++ src/app/core/models/constant-model.ts
@@ -1,0 +1,7 @@
+export class Constant {
+  constructor(
+    public App: string,
+    public Ver: string,
+    public API_URL: string
+  ) {}
+}