+++ src/app/shared/directives/text-emphasis.directive.spec.ts
@@ -1,0 +1,8 @@
+import { TextEmphasisDirective } from './text-emphasis.directive';
+
+describe('TextEmphasisDirective', () => {
+  it('should create an instance', () => {
+    const directive = new TextEmphasisDirective();
+    expect(directive).toBeTruthy();
+  });
+});